# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste API",
        "description": "Teste LIVRO",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Lista todos os livros",
                "description": "Retorna uma coleção paginada de livros",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "Lista de livros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "cod_li": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titulo": {
                                                "type": "string",
                                                "example": "Dom Casmurro"
                                            },
                                            "editora": {
                                                "type": "string",
                                                "example": "Editora Globo"
                                            },
                                            "edicao": {
                                                "type": "string",
                                                "example": 3
                                            },
                                            "ano_publicacao": {
                                                "type": "integer",
                                                "example": 1995
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 59.9
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-06-02T12:34:56Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-06-02T12:34:56Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Cria um novo livro",
                "description": "Cria um livro com dados e relacionamentos (autores e assuntos)",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "editora",
                                    "edicao",
                                    "ano_publicacao",
                                    "preco",
                                    "cod_au",
                                    "cod_as"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "O Cortiço"
                                    },
                                    "editora": {
                                        "type": "string",
                                        "example": "Companhia das Letras"
                                    },
                                    "edicao": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "ano_publicacao": {
                                        "type": "integer",
                                        "example": 1995
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 45.5
                                    },
                                    "cod_au": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "cod_as": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3,
                                            4
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cod_l": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "O Cortiço"
                                        },
                                        "editora": {
                                            "type": "string",
                                            "example": "Companhia das Letras"
                                        },
                                        "edicao": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "ano_publicacao": {
                                            "type": "integer",
                                            "example": 1995
                                        },
                                        "preco": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 45.5
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-05T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-05T12:00:00Z"
                                        },
                                        "authors": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "subjects": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                }
            }
        },
        "/api/books/{cod_l}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Retorna um livro pelo código do livro",
                "operationId": "2bdcb6f4eaf01825d338b92bb530a4ad",
                "parameters": [
                    {
                        "name": "cod_l",
                        "in": "path",
                        "description": "Código do livro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cod_l": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "O Cortiço"
                                        },
                                        "editora": {
                                            "type": "string",
                                            "example": "Companhia das Letras"
                                        },
                                        "edicao": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "ano_publicacao": {
                                            "type": "integer",
                                            "example": 1995
                                        },
                                        "preco": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 45.5
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-05T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-05T12:00:00Z"
                                        },
                                        "authors": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "subjects": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Atualiza um livro existente",
                "operationId": "1059a3d365eba56a978f7049841d0fcf",
                "parameters": [
                    {
                        "name": "cod_l",
                        "in": "path",
                        "description": "Código do livro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "editora",
                                    "edicao",
                                    "ano_publicacao",
                                    "preco",
                                    "cod_au",
                                    "cod_as"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "O Cortiço"
                                    },
                                    "editora": {
                                        "type": "string",
                                        "example": "Companhia das Letras"
                                    },
                                    "edicao": {
                                        "type": "string",
                                        "example": 2
                                    },
                                    "ano_publicacao": {
                                        "type": "integer",
                                        "example": 1995
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "cod_au": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1
                                        ]
                                    },
                                    "cod_as": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Livro atualizado com sucesso"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Remove um livro pelo código do livro",
                "operationId": "624289c1a8f2ec0df83f96c13487d223",
                "parameters": [
                    {
                        "name": "cod_l",
                        "in": "path",
                        "description": "Código do livro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Livro removido com sucesso"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            }
        },
...
    },
...
}